backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pymysql==1.1.0
sqlalchemy==2.0.23
pillow==10.1.0
numpy==1.26.4
tensorflow==2.15.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
pydantic==2.5.0
email-validator==2.3.0
python-dotenv==1.0.0
alembic==1.12.1
pandas==2.1.4

